moves autossl "priming" to a queue to prevent the GCing of the process during normal deploys
